+++ gplsrc/config.h
@@ -18,7 +18,13 @@
  * 
  * Ladybridge Systems can be contacted via the www.openqm.com web site.
  * 
- * START-HISTORY:
+ * ScarletDME Wiki: https://scarlet.deltasoft.com
+ * 
+ * START-HISTORY (ScarletDME):
+ * 27Feb20 gwb Changed integer declarations to be portable across address
+ *             space sizes (32 vs 64 bit)
+ * 
+ * START-HISTORY (OpenQM):
  * 05 Oct 07  2.6-5 Added PDUMP parameter.
  * 20 Aug 07  2.6-0 Added CMDSTACK parameter.
  * 01 Jul 07  2.5-7 Extensive changes for PDA merge.
@@ -53,27 +59,27 @@
 
 struct CONFIG
  {
-  short int max_users;             /* User limit */
+  int16_t max_users;             /* User limit */
   char sysdir[MAX_PATHNAME_LEN+1];
-  short int cmdstack;              /* CMDSTACK: Command stack depth */
+  int16_t cmdstack;              /* CMDSTACK: Command stack depth */
   bool deadlock;                   /* DEADLOCK: Trap deadlocks */
-  unsigned short int debug;        /* DEBUG:    Controls debug features */
+  u_int16_t debug;        /* DEBUG:    Controls debug features */
   int errlog;                      /* ERRLOG:   Max errlog size in bytes, 0 if disabled */
-  short int fds_limit;             /* FDS */
-  short int fixusers_base;         /* FIXUSERS: First user number and... */
-  short int fixusers_range;        /*          ...Number of users */
-  short int jnlmode;               /* JNLMODE:  Journalling mode */
+  int16_t fds_limit;             /* FDS */
+  int16_t fixusers_base;         /* FIXUSERS: First user number and... */
+  int16_t fixusers_range;        /*          ...Number of users */
+  int16_t jnlmode;               /* JNLMODE:  Journalling mode */
   char jnldir[MAX_PATHNAME_LEN+1]; /* JNLDIR:   Journal file directory */
-  short int maxidlen;              /* MAXIDLEN: Maximum record id length */
-  short int netfiles;              /* NETFILES:
+  int16_t maxidlen;              /* MAXIDLEN: Maximum record id length */
+  int16_t netfiles;              /* NETFILES:
                                       0x0001    Allow outgoing NFS
                                       0x0002    Allow incoming QMNet   */
-  short int numfiles;              /* NUMFILES: Maximum number of files open */
-  short int numlocks;              /* NUMLOCKS: Maximum number of record locks */
-  short int pdump;                 /* PDUMP:    PDUMP mode flags */
-  short int portmap_base_port;     /* PORTMAP: First port number ... */
-  short int portmap_base_user;     /*          ...First user number... */
-  short int portmap_range;         /*          ...Number of ports/users */
+  int16_t numfiles;              /* NUMFILES: Maximum number of files open */
+  int16_t numlocks;              /* NUMLOCKS: Maximum number of record locks */
+  int16_t pdump;                 /* PDUMP:    PDUMP mode flags */
+  int16_t portmap_base_port;     /* PORTMAP: First port number ... */
+  int16_t portmap_base_user;     /*          ...First user number... */
+  int16_t portmap_range;         /*          ...Number of ports/users */
   char startup[80+1];              /* STARTUP: Startup command */
  };
 
@@ -86,32 +92,32 @@ struct PCFG
   unsigned int codepage;                /* CODEPAGE: Set console codepage */
   char dumpdir[MAX_PATHNAME_LEN+1];     /* DUMPDIR:  Directory for process dump files */
   bool exclrem;                         /* EXCLREM:  Exclude remote files from ACCOUNT.SAVE? */
-  short int filerule;                   /* FILERULE: Rules for special filename formats */
+  int16_t filerule;                     /* FILERULE: Rules for special filename formats */
   double fltdiff;                       /* FLTDIFF : Wide zero for float comparisons */
-  short int fsync;                      /* FSYNC:    Controls when to do fsync() */
+  int16_t fsync;                        /* FSYNC:    Controls when to do fsync() */
   bool gdi;                             /* GDI:      Default to GDI print API? */
-  short int grpsize;                    /* GRPSIZE:  Default group size (1-8) */
-  short int intprec;                    /* INTPREC:  Precision for INT() etc */
-  short int lptrhigh;                   /* LPTRHIGH: Default printer lines */
-  short int lptrwide;                   /* LPTRWIDE: Default printer width */
+  int16_t grpsize;                      /* GRPSIZE:  Default group size (1-8) */
+  int16_t intprec;                      /* INTPREC:  Precision for INT() etc */
+  int16_t lptrhigh;                     /* LPTRHIGH: Default printer lines */
+  int16_t lptrwide;                     /* LPTRWIDE: Default printer width */
   int maxcall;                          /* MAXCALL:  Maximum call depth */
   bool must_lock;                       /* MUSTLOCK: Enforce locking rules */
-  short int objects;                    /* OBJECTS:  Max loaded objects */
-  long int objmem;                      /* OBJMEM:   Object size limit, zero = none */
-  short int qmclient_mode;              /* QMCLIENT: 0 = any, 1 = no open/exec, 2 = restricted call */
-  short int reccache;                   /* RECCACHE: Record cache size */
+  int16_t objects;                      /* OBJECTS:  Max loaded objects */
+  int32_t objmem;                       /* OBJMEM:   Object size limit, zero = none */
+  int16_t qmclient_mode;                /* QMCLIENT: 0 = any, 1 = no open/exec, 2 = restricted call */
+  int16_t reccache;                     /* RECCACHE: Record cache size */
   bool ringwait;                        /* RINGWAIT: Wait if ring buffer full */
   bool safedir;                         /* SAFEDIR:  Use careful update on dir file write */
   char sh[MAX_SH_CMD_LEN+1];            /* SH:       Command to run interactive shell */
   char sh1[MAX_SH_CMD_LEN+1];           /* SH1:      Command to run single shell */
-  long int sortmem;                     /* SORTMEM: Limit on in-memory sort size */
-  short int sortmrg;                    /* SORTMRG: Number of files in merge */
+  int32_t sortmem;                      /* SORTMEM: Limit on in-memory sort size */
+  int16_t sortmrg;                      /* SORTMRG: Number of files in merge */
   char sortworkdir[MAX_PATHNAME_LEN+1]; /* SORTWORK */
   char spooler[MAX_PATHNAME_LEN+1];     /* SPOOLER: Non-default spooler */
   char tempdir[MAX_PATHNAME_LEN+1];     /* TEMPDIR */
   char terminfodir[MAX_PATHNAME_LEN+1]; /* TERMINFO */
   bool txchar;                          /* TXCHAR */
-  short int yearbase;                   /* YEARBASE: Two digit year start */
+  int16_t yearbase;                   /* YEARBASE: Two digit year start */
  };
 
 Public struct PCFG pcfg;
