+++ gplsrc/linuxlb.h
@@ -18,7 +18,13 @@
  * 
  * Ladybridge Systems can be contacted via the www.openqm.com web site.
  * 
- * START-HISTORY:
+ * ScarletDME Wiki: https://scarlet.deltasoft.com
+ * 
+ * START-HISTORY (ScarletDME):
+ * 27Feb20 gwb Changed integer declarations to be portable across address
+ *             space sizes (32 vs 64 bit)
+ * 
+ * START-HISTORY (OpenQM):
  * 01 Jul 07  2.5-7 Extensive change for PDA merge.
  * 16 Sep 04  2.0-1 OpenQM launch. Earlier history details suppressed.
  * END-HISTORY
@@ -35,14 +41,14 @@
 
 /* Inline macros */
 
-#define max(a,b)    (((a) > (b)) ? (a) : (b))
-#define min(a,b)    (((a) < (b)) ? (a) : (b))
+#define max(a, b) (((a) > (b)) ? (a) : (b))
+#define min(a, b) (((a) < (b)) ? (a) : (b))
 
 /* Simple substitutes */
 
-#define chsize(fd,bytes) ftruncate(fd,bytes)
+#define chsize(fd, bytes) ftruncate(fd, bytes)
 #define GetCurrentProcessId() getpid()
-#define stricmp(a,b) strcasecmp(a,b)
+#define stricmp(a, b) strcasecmp(a, b)
 
 #define chsize64 chsize
 
@@ -50,12 +56,12 @@
 
 int64 filelength64(int fd);
 #define filelength(f) (int)filelength64(f)
-bool GetUserName(char * name, unsigned long int * bytes);
-char * itoa(int value, char * string, int radix);
-char * Ltoa(long int value, char * string, int radix);
-char * qmrealpath(char * inpath, char * outpath);
-void Sleep(long int n);
-void strrep(char * s, char old, char new);
+bool GetUserName(char* name, u_int32_t* bytes);
+char* itoa(int value, char* string, int radix);
+char* Ltoa(int32_t value, char* string, int radix);
+char* qmrealpath(char* inpath, char* outpath);
+void Sleep(int32_t n);
+void strrep(char* s, char old, char new);
 
 #endif
